backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
PyJWT
httpx
bcrypt
ldap3
openpyxl
